backend/requirements.txt
django
djangorestframework
openai
python-dotenv
pytest-django
psycopg2-binary
django-cors-headers
